With previous version, call functionality was duplicated in Male and Female class.
So I have decided to create a base class called Mobile and inherit both Male and Female class.
So now there is only one call method in Mobile class.
Observer we have changed only Male and Female, but if you know run the same driving code , it will still give you the same output as before.
That means we are not breaking any existing code by creating Mobile and inheriting in Male and Female class
So we made sure to avoid duplicated code of Male and Female without disturbing our existing driving code